Script_generation/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.11.0
httpx==0.27.0
requests==2.31.0
python-dotenv
